Consolidate config require/forwarded helpers into util/config.py
run_all_analyses.py's require(cfg, dotted) generalised two existing
lookup-and-raise implementations (g1_scan_validation._require and
scan_orchestrator.require_keys) without replacing either, and its
column_kwargs/hex_kwargs were the same "forward only the keys the config
sets" function written twice, differing by a no-op float() cast.

- require and forwarded now live in util/config.py, next to deep_get.
- g1_scan_validation._require is deleted; its two call sites use require
  directly, same messages, no behaviour change (params was already in
  scope).
- scan_orchestrator.require_keys stays separate, with a comment: it runs
  in a worker process/pool, so it must raise KeyError, not ClickException.
- Their tests move to test_config_flows.py alongside the other util/config
  primitives.
